Remove duplicate FreshAtomicIdSource
