+++ mimir/src/test/scala/quasar/mimir/Evaluator.scala
@@ -16,7 +16,6 @@
 
 package quasar.mimir
 
-import quasar.blueeyes._
 import quasar.precog.common._
 import quasar.yggdrasil.bytecode._
 import quasar.yggdrasil._
